Decorate docs refactor
