+++ server/views/pages/newGeneratedDocumentVersion.njk
@@ -1,6 +1,6 @@
 {% extends "../partials/layout.njk" %}
 
-{% set pageTitle = applicationName + " - Book a recall - Upload a new document" %}
+{% set pageTitle = applicationName + " - Create a new " %}
 
 {% block beforeContent %}
     {{ govukBackLink({
